+++ libs/common/src/constants/tfm/tfm-deal-cancellation-status.ts
