+++ libs/common/src/types/tfm/tfm-deal-cancellation-status.ts
